requirements.txt
Flask==2.1.2
Flask-Cors==3.0.10
python-dotenv==0.20.0
